make fast guts load optional with -fastguts
In the C++ standard library, the data structure std::queue
does not automatically release memory after removing all
elements because std::queue is based on a container that
retains allocated memory even when it becomes empty.

We made the fast guts load implementation optional with
the `-fastguts` flag. It offers nearly double the loading
speed but also doubles memory consumption.